add input wrapper
